do not map, change properties or set focus on the currently active content
Otherwise, the window manager crashes when windows are destroyed in rapid
succession. For example, an application may open window A and B. When A is
active it quits; it first destroys B and then A. The window manager is first
notified of the removal of B and calls panelenter(), which maps the active
window, which is still A. When the server receives this map request A is
already destroyed.