handle BadWindow and BadAtom errors
The BadWindow errors are unavoidable because other programs may close two
windows in a very short time, but the X server notifies irwm one at time.
Dealing with the closure of the first window may make the second window active,
while it is already closed.

An example is gedit when closing without saving. When the "don't save" button
in the confirmation window is pressed, gedit closes the confirmation window and
its main window. The X server first notifies irwm of the first closure. This
causes irwm to delete the panel of the confirmation window and enter the panel
below it, which contains the main gedit window. This causes mapping, setting
focus and other operations on a window that is already closed, resulting in
BadWindow errors. This situation cannot be avoided since the X server has not
informed irwm of the closure of the main gedit window yet.

More generally, programs may close windows at arbitrary times. The X server
notifies irwm one at time. The closure of one may result in making another
active, which requires operating on an window which is already closed.

Windows with the override_redirect may be closed between the time they are
mapped or configured and irwm finishes moving them. This is why errors from
getting windows attributes need to be handled.

Also unavoidable are BadAtom errors. Other programs may send whatever
_NET_WM_STATE messages they want, including data that do not contain valid atom
identifiers. This results in BadAtom errors that need to be handled.