randomize the position of override_redirect windows
Sometimes programs create menus that are taller than the screen. Just lining
them with the upper border of the screen hides their last entries. With this
commit, every time they are created they are moved to a random position. A
preference is given to lining them with the upper or lower border since
otherwise the first and last entries are unlikely to be shown.